+++ eslint.config.js
@@ -1,38 +1,39 @@
 /* eslint-disable */ //** remove */
-const { FlatCompat } = require("@eslint/eslintrc");
+const { FlatCompat } = require('@eslint/eslintrc')
 
-const path = require("path");
-const { fileURLToPath } = require("url");
+const path = require('path')
+const { fileURLToPath } = require('url')
 
 const compat = new FlatCompat({
   baseDirectory: path.dirname(
-    fileURLToPath(require("url").pathToFileURL(__filename).toString())
+    fileURLToPath(require('url').pathToFileURL(__filename).toString())
   ),
-});
+})
 
 module.exports = [
   ...compat.extends(
-    "next",
-    "next/core-web-vitals",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:prettier/recommended"
+    'next',
+    'next/core-web-vitals',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended'
   ),
   {
-    files: ["**/*.{js,jsx,ts,tsx}"],
+    files: ['**/*.{js,jsx,ts,tsx}'],
     // parser: '@typescript-eslint/parser',
     // parserOptions: {
     //     ecmaVersion: 2023,
     //     sourceType: 'module',
     // },
     rules: {
-      "prettier/prettier": [
-        "error",
+      'prettier/prettier': [
+        'error',
         {
           singleQuote: true,
-          endOfLine: "auto",
+          endOfLine: 'auto',
+          semi: false,
         },
       ],
-      // Other rules...
+      '@typescript-eslint/no-var-requires': 'off',
     },
-  }
-];
+  },
+]