+++ src/app/[locale]/(auth)/(center)/sign-in/[[...sign-in]]/page.tsx
@@ -1,22 +1,22 @@
-import { SignIn } from '@clerk/nextjs';
-import { getTranslations } from 'next-intl/server';
+import { SignIn } from '@clerk/nextjs'
+import { getTranslations } from 'next-intl/server'
 
-import { getI18nPath } from '@/utils/Helpers';
+import { getI18nPath } from '@/utils/Helpers'
 
 export async function generateMetadata(props: { params: { locale: string } }) {
   const t = await getTranslations({
     locale: props.params.locale,
-    namespace: 'SignIn',
-  });
+    namespace: 'Login',
+  })
 
   return {
     title: t('meta_title'),
     description: t('meta_description'),
-  };
+  }
 }
 
 const SignInPage = (props: { params: { locale: string } }) => (
   <SignIn path={getI18nPath('/sign-in', props.params.locale)} />
-);
+)
 
-export default SignInPage;
+export default SignInPage