+++ src/app/[locale]/(auth)/dashboard/layout.tsx
@@ -2,8 +2,8 @@
 
 // import LocaleSwitcher from '@/components/LocaleSwitcher';
 // import { LogOutButton } from '@/components/LogOutButton';
-import { BaseTemplate } from '@/templates/BaseTemplate';
-import NavMain from '../../../../components/tailwind-ui/application-ui/navigation/navbars/dark_with_quick_action';
+import { BaseTemplate } from '@/templates/BaseTemplate'
+import NavMain from '../../../../components/tailwind-ui/application-ui/navigation/navbars/dark_with_quick_action'
 
 export default function DashboardLayout(props: { children: React.ReactNode }) {
   // const t = useTranslations('DashboardLayout');
@@ -12,5 +12,5 @@ export default function DashboardLayout(props: { children: React.ReactNode }) {
     <BaseTemplate leftNav={<NavMain />} rightNav={<NavMain />}>
       {props.children}
     </BaseTemplate>
-  );
+  )
 }