+++ src/components/icons/ReloadIcon.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import Icon from '@mdi/react'
+import { mdiReload } from '@mdi/js'
+
+const ReloadIcon = () => (
+  <div className="ml-px -mr-px">
+    <Icon path={mdiReload} size={1} color="white" />
+  </div>
+)
+
+export default ReloadIcon