+++ src/components/tailwind-ui/application-ui/navigation/navbars/dark_with_quick_action.jsx
@@ -1,29 +1,29 @@
-import { AppConfig } from '@/utils/AppConfig';
-import { Fragment } from 'react';
-import { Disclosure, Menu, Transition } from '@headlessui/react';
-import { Bars3Icon, BellIcon, XMarkIcon } from '@heroicons/react/24/outline';
-import { BR } from 'country-flag-icons/react/3x2';
+import { AppConfig } from '@/utils/AppConfig'
+import { Fragment } from 'react'
+import { Disclosure, Menu, Transition } from '@headlessui/react'
+import { Bars3Icon, BellIcon, XMarkIcon } from '@heroicons/react/24/outline'
+import { BR } from 'country-flag-icons/react/3x2'
 
 const user = {
   name: 'Tom Cook',
   email: 'tom@example.com',
   imageUrl:
     'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-};
+}
 const navigation = [
   { name: 'Dashboard', href: '#', current: true },
   { name: 'Team', href: '#', current: false },
   { name: 'Projects', href: '#', current: false },
   { name: 'Calendar', href: '#', current: false },
-];
+]
 const userNavigation = [
   { name: 'Your Profile', href: '#' },
   { name: 'Settings', href: '#' },
   { name: 'Sign out', href: '#' },
-];
+]
 
 function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
+  return classes.filter(Boolean).join(' ')
 }
 
 export default function NavMain() {
@@ -48,7 +48,7 @@ export default function NavMain() {
                 <div>
                   <img
                     className="mx-auto my-3"
-                    src="/assets/images/logo/casa-dourada-dark.png"
+                    src="/assets/images/logo/casa-dourada-light.png"
                     alt={AppConfig.name}
                     width={108}
                     height={40}
@@ -199,5 +199,5 @@ export default function NavMain() {
         </>
       )}
     </Disclosure>
-  );
+  )
 }