+++ src/services/strapiService.js
@@ -1,13 +1,11 @@
-import axios from 'axios';
+import axios from 'axios'
 
-const STRAPI_API_URL = process.env.STRAPI_API_URL;
-const TOKEN = process.env.TOKEN;
+const STRAPI_API_URL = process.env.STRAPI_API_URL
+const TOKEN = process.env.STRAPI_TOKEN
 
 const strapiService = {
   async getPostByCategoryAndSlug(slug, category) {
-    const url = `${STRAPI_API_URL}posts?filters[category][slug][$eq]=${category}&filters[slug][$eq]=${slug}`;
-
-    console.log(`Fetching URL: ${url}`);
+    const url = `${STRAPI_API_URL}posts?filters[category][slug][$eq]=${category}&filters[slug][$eq]=${slug}`
 
     try {
       const response = await axios.get(url, {
@@ -15,24 +13,48 @@ const strapiService = {
           Authorization: `Bearer ${TOKEN}`,
           'Content-Type': 'application/json',
         },
-      });
-
-      console.log('API Response:', response.data);
+      })
 
       if (
         response.data &&
-        Array.isArray(response.data.data) && // Changed from response.data to response.data.data
+        Array.isArray(response.data.data) &&
         response.data.data.length > 0
       ) {
-        return response.data.data[0];
+        return response.data.data[0].attributes
       } else {
-        throw new Error('Post not found');
+        throw new Error('Post not found')
       }
     } catch (error) {
-      console.error(`Error fetching post from ${url}:`, error);
-      throw error;
+      console.error(`Error fetching post from ${url}:`, error)
+      throw error
     }
   },
-};
 
-export default strapiService;
+  async getAllPosts() {
+    const url = `${STRAPI_API_URL}posts`
+
+    try {
+      const response = await axios.get(url, {
+        headers: {
+          Authorization: `Bearer ${TOKEN}`,
+          'Content-Type': 'application/json',
+        },
+      })
+
+      if (
+        response.data &&
+        Array.isArray(response.data.data) &&
+        response.data.data.length > 0
+      ) {
+        return response.data.data.map((post) => post.attributes)
+      } else {
+        throw new Error('No posts found')
+      }
+    } catch (error) {
+      console.error(`Error fetching posts from ${url}:`, error)
+      throw error
+    }
+  },
+}
+
+export default strapiService