+++ src/utils/AppConfig.ts
@@ -1,11 +1,11 @@
-import type { LocalePrefix } from 'node_modules/next-intl/dist/types/src/routing/types';
+import type { LocalePrefix } from 'node_modules/next-intl/dist/types/src/routing/types'
 
-const localePrefix: LocalePrefix = 'as-needed';
+const localePrefix: LocalePrefix = 'as-needed'
 
 // FIXME: Update this configuration file based on your project information
 export const AppConfig = {
   name: 'Casa Dourada',
-  locales: ['en', 'fr'],
-  defaultLocale: 'en',
+  locales: ['br', 'en'],
+  defaultLocale: 'br',
   localePrefix,
-};
+}